requirements.txt
faster-whisper==1.1.1
dotenv==0.9.9
pydub==0.25.1
pyannote.audio==3.3.2